requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-postprocessor-cohere-rerank
llama-index-vector-stores-qdrant
qdrant-client
fastapi
openai
pydantic_settings
docling
markitdown
python-dotenv
uvicorn